unify cmdK
