+++ src/features/SearchBox/SearchResults.tsx
@@ -7,16 +7,17 @@ import { useNavigate } from 'react-router';
 import { useDebounce } from 'usehooks-ts';
 
 import { LoadingIndicator } from '../../components/LoadingIndicator';
-import { AiLight } from '../../icons/__generated';
 import { coLinksPaths } from '../../routes/paths';
 import { Avatar, Flex, Text } from '../../ui';
 import { SkillTag } from '../colinks/SkillTag';
 
+import { FarcasterUserResult } from './FarcasterUserResult';
+import { fetchFarcasterUserResults } from './fetchFarcasterUserResults';
+import { fetchPartyProfileResults } from './fetchPartyProfileResults';
 import { fetchPostSearchResults } from './fetchPostSearchResults';
-import { fetchSearchResults } from './fetchSearchResults';
-import { fetchSimilarityResults } from './fetchSimilarityResults';
 import { PeopleResult } from './PeopleResult';
 import { SearchBoxPages } from './SearchBoxPages';
+import { useGiveSkills } from './useGiveSkills';
 
 const QUERY_KEY_SEARCH = 'searchBoxQuery';
 const QUERY_KEY_SEARCH_POSTS = 'searchBoxQueryPosts';
@@ -32,38 +33,45 @@ export const SearchResults = ({
   const debouncedSearch = useDebounce(search, 300);
   const navigate = useNavigate();
 
-  const searchIsEmpty = search === '';
+  const { data: partyProfileResults, isFetching: partyProfileFetching } =
+    useQuery(
+      [QUERY_KEY_SEARCH, 'partyProfiles', JSON.stringify(debouncedSearch)],
+      async () =>
+        await fetchPartyProfileResults({
+          search: debouncedSearch,
+        }),
+      {
+        enabled: debouncedSearch !== '',
+        staleTime: Infinity,
+        retry: false,
+      }
+    );
 
-  const { data: similarityResults, isFetching: similarityFetching } = useQuery(
-    [QUERY_KEY_SEARCH, 'similarity', JSON.stringify(debouncedSearch)],
-    async () =>
-      await fetchSimilarityResults({ search: JSON.stringify(debouncedSearch) }),
-    {
-      enabled: debouncedSearch !== '',
-      staleTime: Infinity,
-      retry: false,
-    }
-  );
+  const { data: farcasterUserResults, isFetching: farcasterUserFetching } =
+    useQuery(
+      [QUERY_KEY_SEARCH, 'farcasterUsers', JSON.stringify(debouncedSearch)],
+      async () =>
+        await fetchFarcasterUserResults({
+          search: debouncedSearch,
+        }),
+      {
+        enabled: debouncedSearch !== '',
+        staleTime: Infinity,
+        retry: false,
+      }
+    );
 
-  const { data: results, isLoading } = useQuery(
-    [QUERY_KEY_SEARCH, JSON.stringify(debouncedSearch)],
-    () =>
-      fetchSearchResults({
-        where: {
-          name: debouncedSearch
-            ? { _ilike: `%${debouncedSearch}%` }
-            : undefined,
-        },
-        skillsWhere: {
-          name: debouncedSearch
-            ? { _ilike: `%${debouncedSearch}%` }
-            : undefined,
-        },
-      }),
-    {
-      // staleTime: Infinity,
-    }
-  );
+  const giveSkills = useGiveSkills(debouncedSearch);
+
+  // TODO: dedupe farcaster results if they are in profile results
+
+  const onSelectAddress = (address: string) => {
+    navigate(coLinksPaths.profileGive(address));
+  };
+
+  const onSelectSkill = (name: string) => {
+    navigate(coLinksPaths.giveSkill(name));
+  };
 
   const { data: postResults } = useQuery(
     [QUERY_KEY_SEARCH_POSTS, JSON.stringify(debouncedSearch)],
@@ -75,14 +83,6 @@ export const SearchResults = ({
       // staleTime: Infinity,
     }
   );
-
-  const onSelectProfile = (address: string) => {
-    navigate(coLinksPaths.profileGive(address));
-  };
-
-  const onSelectInterest = (name: string) => {
-    navigate(coLinksPaths.exploreSkill(name));
-  };
 
   const onSelectPost = (id: string) => {
     navigate(coLinksPaths.post(id));
@@ -99,9 +99,7 @@ export const SearchResults = ({
       <Command.List>
         <Command.Empty>No results found.</Command.Empty>
         <SearchBoxPages search={search} />
-        {/* || */}
-        {/* (similarityFetching  */}
-        {isLoading && (
+        {(partyProfileFetching || farcasterUserFetching) && (
           <Command.Loading>
             <Flex column css={{ width: '100%', alignItems: 'center' }}>
               <Flex css={{ alignItems: 'center', gap: '$md' }}>
@@ -114,163 +112,134 @@ export const SearchResults = ({
           </Command.Loading>
         )}
         <>
-          <Command.Group heading={'People'}>
-            {results?.profiles_public?.map(profile => (
+          {(giveSkills?.length ?? 0) > 0 && (
+            <Command.Group heading={'Skills'}>
+              {giveSkills
+                ?.filter(skill => skill.skill)
+                .map(skill => (
+                  <Command.Item
+                    key={skill.skill}
+                    value={skill.skill}
+                    onSelect={name => {
+                      onSelectSkill(name);
+                      setPopoverOpen(false);
+                    }}
+                  >
+                    <SkillTag
+                      give={true}
+                      skill={skill.skill}
+                      count={skill.gives}
+                      size={'medium'}
+                    />
+                  </Command.Item>
+                ))}
+            </Command.Group>
+          )}
+          {(partyProfileResults?.length ?? 0) > 0 && (
+            <Command.Group heading={'CoLinks Members'}>
+              {partyProfileResults
+                ?.filter(p => p.name && p.id)
+                .map(profile => (
+                  <Command.Item
+                    key={'ppr_' + profile.id}
+                    value={profile.address}
+                    onSelect={address => {
+                      onSelectAddress(address);
+                      setPopoverOpen(false);
+                    }}
+                  >
+                    <PeopleResult profile={profile} />
+                  </Command.Item>
+                ))}
+            </Command.Group>
+          )}
+          {(farcasterUserResults?.length ?? 0) > 0 && (
+            <Command.Group heading={'Farcaster Users'}>
+              {farcasterUserResults
+                ?.filter(u => u.fname && u.address)
+                .map(user => (
+                  <Command.Item
+                    key={user.fname}
+                    value={user.address}
+                    onSelect={address => {
+                      onSelectAddress(address);
+                      setPopoverOpen(false);
+                    }}
+                  >
+                    <FarcasterUserResult user={user} />
+                  </Command.Item>
+                ))}
+            </Command.Group>
+          )}
+        </>
+        {debouncedSearch !== '' && (
+          <Command.Group heading={'Posts'}>
+            {postResults?.posts?.map(post => (
               <Command.Item
-                key={profile.id}
-                value={profile.address}
-                onSelect={address => {
-                  onSelectProfile(address);
+                key={`${post.activity?.id}`}
+                value={`${post.activity?.id}`}
+                onSelect={id => {
+                  onSelectPost(id);
                   setPopoverOpen(false);
                 }}
               >
-                <PeopleResult profile={profile} />
-              </Command.Item>
-            ))}
-          </Command.Group>
-          <Command.Group heading={'Interests'}>
-            {results?.interests?.map(interest => (
-              <Command.Item
-                key={interest.name}
-                value={interest.name}
-                onSelect={name => {
-                  onSelectInterest(name);
-                  setPopoverOpen(false);
-                }}
-              >
-                <SkillTag
-                  skill={interest.name}
-                  count={interest.count}
-                  size={'medium'}
+                <Avatar
+                  size="xs"
+                  name={post.profile_public?.name}
+                  path={post.profile_public?.avatar}
                 />
-              </Command.Item>
-            ))}
-          </Command.Group>
-          {debouncedSearch !== '' && (
-            <Command.Group heading={'Posts'}>
-              {postResults?.posts?.map(post => (
-                <Command.Item
-                  key={`${post.activity?.id}`}
-                  value={`${post.activity?.id}`}
-                  onSelect={id => {
-                    onSelectPost(id);
-                    setPopoverOpen(false);
+                <Flex
+                  column
+                  css={{
+                    width: '100%',
                   }}
                 >
-                  <Avatar
-                    size="xs"
-                    name={post.profile_public?.name}
-                    path={post.profile_public?.avatar}
-                  />
                   <Flex
-                    column
                     css={{
+                      justifyContent: 'space-between',
+                      alignItems: 'center',
                       width: '100%',
                     }}
                   >
                     <Flex
                       css={{
-                        justifyContent: 'space-between',
+                        gap: '$sm',
                         alignItems: 'center',
-                        width: '100%',
+                        flexGrow: 1,
                       }}
                     >
-                      <Flex
-                        css={{
-                          gap: '$sm',
-                          alignItems: 'center',
-                          flexGrow: 1,
-                        }}
-                      >
-                        <Text semibold size={'xs'}>
-                          {post.profile_public?.name}
-                        </Text>
-                      </Flex>
-                      <Flex
-                        css={{
-                          justifyContent: 'flex-end',
-                          flexShrink: 0,
-                        }}
-                      >
-                        <Text
-                          size="xs"
-                          css={{
-                            color: '$neutral',
-                            ml: '$md',
-                          }}
-                        >
-                          {DateTime.fromISO(post.created_at).toRelative({
-                            style: 'narrow',
-                          })}
-                        </Text>
-                      </Flex>
+                      <Text semibold size={'xs'}>
+                        {post.profile_public?.name}
+                      </Text>
                     </Flex>
-                    <Text size={'xs'}>
-                      {post.description.slice(0, 100)}
-                      {post.description.length > 100 ? '...' : ''}
-                    </Text>
-                  </Flex>
-                </Command.Item>
-              ))}
-            </Command.Group>
-          )}
-          {!searchIsEmpty && (
-            <Command.Group
-              heading={
-                <Flex
-                  css={{
-                    gap: '$sm',
-                    alignItems: 'center',
-                  }}
-                >
-                  <AiLight
-                    size={'lg'}
-                    nostroke
-                    css={{ '*': { fill: '$text' } }}
-                  />
-                  Suggested People
-                </Flex>
-              }
-            >
-              {similarityFetching && (
-                <Command.Item
-                  key={'ai-loading'}
-                  value="searchSimilarProfiles"
-                  data-disabled={true}
-                  data-selectable={false}
-                >
-                  <Text size={'xs'} color={'neutral'} css={{}}>
-                    <AiLight
-                      size={'lg'}
-                      nostroke
-                      css={{ mr: '$xs', '*': { fill: '$secondaryText' } }}
-                    />
-                    Searching for similar profiles ...{' '}
-                  </Text>
-                </Command.Item>
-              )}
-              {similarityResults?.map(
-                res =>
-                  res &&
-                  res.profile_public && (
-                    <Command.Item
-                      key={res.profile_public?.id}
-                      value={res.profile_public?.address}
-                      onSelect={address => {
-                        onSelectProfile(address);
-                        setPopoverOpen(false);
+                    <Flex
+                      css={{
+                        justifyContent: 'flex-end',
+                        flexShrink: 0,
                       }}
                     >
-                      <PeopleResult
-                        profile={res.profile_public}
-                        score={res.similarity}
-                      />
-                    </Command.Item>
-                  )
-              )}
-            </Command.Group>
-          )}
-        </>
+                      <Text
+                        size="xs"
+                        css={{
+                          color: '$neutral',
+                          ml: '$md',
+                        }}
+                      >
+                        {DateTime.fromISO(post.created_at).toRelative({
+                          style: 'narrow',
+                        })}
+                      </Text>
+                    </Flex>
+                  </Flex>
+                  <Text size={'xs'}>
+                    {post.description.slice(0, 100)}
+                    {post.description.length > 100 ? '...' : ''}
+                  </Text>
+                </Flex>
+              </Command.Item>
+            ))}
+          </Command.Group>
+        )}
       </Command.List>
     </>
   );
