+++ src/features/colinks/SkillTag.tsx
@@ -3,16 +3,18 @@ import { useNavigate } from 'react-router-dom';
 import { coLinksPaths } from '../../routes/paths';
 import { CSS, skillTextStyle } from '../../stitches.config';
 import { Flex, Text } from '../../ui';
-import { Users } from 'icons/__generated';
+import { GemCoOutline, Users } from 'icons/__generated';
 
 export const SkillTag = ({
   skill,
+  give = false,
   count,
   size = 'medium',
   css,
   active,
 }: {
   skill: string;
+  give?: boolean;
   count?: number;
   // size?: ComponentProps<typeof Text>['size'];
   size?: 'large' | 'medium' | 'small';
@@ -57,7 +59,11 @@ export const SkillTag = ({
             },
           }}
         >
-          <Users size={size === 'large' ? 'lg' : undefined} />
+          {give ? (
+            <GemCoOutline fa size={size === 'large' ? 'lg' : undefined} />
+          ) : (
+            <Users size={size === 'large' ? 'lg' : undefined} />
+          )}
           {count}
         </Flex>
       )}
